api/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
spacy
ftfy
unidecode
python-dotenv
google-genai